==> scripts/01-data-prep/05_Socialization_Data.R <==
#-----------------Socialization Data Pre-Processing Script----------------------
#-Author: A. Jordan Nafa----------------------------Created: September 9, 2021-#
#-R Version: 4.1.2----------------------------------Last Revised: May 31, 2022-#

# Load the necessary libraries----
pacman::p_load(
  "sjlabelled",
  "tidyverse",
  "data.table",
  "dtplyr",
  "collapse",
  "countrycode",
  install = FALSE
)

# Read the IVS Model Data from script 01
ivs_df_sub <- read_rds("data/IVS/IVS_Subset.rds")

# Read in the contextual variables from script 02
context_df <- read_rds("data/Context_Data.rds")

#------------------------------------------------------------------------------#
#--------------------Socialization Data, Five Year Window-----------------------
#------------------------------------------------------------------------------#

ivs_soc_df_five <- ivs_df_sub %>%
  # Subset the identifiers and respondent's birth year
  select(country:resp_id, age:birth_year) %>%
  # Calculate a respondent's socialization period
  mutate(
    across(
      birth_year,
      list(
        start = ~ .x + 16,
        end = ~ .x + 21
      ),
      .names = "soc_{.fn}"
    )
  ) %>%
  # Drop missing values in birth year
  drop_na(birth_year) %>%
  # Group the data by respondent
  group_by(resp_id) %>%
  # Nest the data
  nest(soc_period = c(soc_start, soc_end)) %>%
  # Expand each respondent's socialization period
  mutate(soc_period = map(
    soc_period, ~ seq(.x$soc_start, .x$soc_end, 1)
  )) %>%
  # Unnest the data
  unnest(cols = c(soc_period)) %>%
  # Generate a vector for socialization range
  mutate(
    soc_range = str_c(
      min(soc_period),
      max(soc_period),
      sep = "-"
    )) %>%
  # Ungroup the data
  ungroup() %>%
  # Generate Socialization Age
  mutate(soc_age = soc_period - birth_year) %>%
  # Filter Socialization Age <= Current Age
  filter(soc_age <= age) %>%
  # Merge in the V-Dem Socialization Variables
  left_join(
    context_df %>%
      select(
        country:vdem_iso,
        soc_pctfemleg5 = pct_female_leg,
        soc_libdem5 = libdem,
        soc_polyarchy5 = polyarchy,
        soc_legislature5 = legislature,
        soc_election_year5 = election_year,
        soc_regime_type5 = regime_type
      ),
    by = c("country", "soc_period" = "year")
  )

# Recovering the ISO3 and V-Dem IDs
soc_ids_five <- ivs_soc_df_five %>%
  # Group the data by respondent
  group_by(resp_id) %>%
  # Collapse the IDs by respondent
  summarise(across(
    iso_code:vdem_iso,
    ~ paste0(unique(.x), collapse = ', '),
    .names = "soc_{.col}5"
  )) %>%
  # Ungroup the data
  ungroup()

ivs_soc_df_five <- ivs_soc_df_five %>% 
  # Group the data by identifiers
  group_by(country, year, resp_id) %>%
  # Collapse the data by respondent
  summarise(
    # Respondent's Five-Year Socialization Window
    window_5y = unique(soc_range),
    # Take the maximum value of Female HoG and Regime Type
    across(
      c(soc_legislature5:soc_election_year5, soc_regime_type5),
      ~ max(as.integer(.x), na.rm = T)
    ),
    # Take the average of continuous and count variables
    across(
      soc_pctfemleg5:soc_polyarchy5,
      ~ mean(.x, na.rm = T)
    )
  ) %>%
  # Ungroup the data
  ungroup() %>%
  # Fix any non-finite values generated by aggregation
  mutate(across(
    starts_with("soc_"),
    ~ case_when(
      is.infinite(.x) | is.nan(.x) ~ NA_real_,
      TRUE ~ .x
    )
  )) %>% 
  # Ensure no observations were duplicated
  distinct(resp_id, .keep_all = TRUE)

#------------------------------------------------------------------------------#
#---------------------Socialization Data, Ten Year Window-----------------------
#------------------------------------------------------------------------------#

ivs_soc_df_ten <- ivs_df_sub %>%
  # Subset the identifiers and respondent's birth year
  select(country:resp_id, age:birth_year) %>%
  # Calculalte a respondent's socialization period
  mutate(
    across(
      birth_year,
      list(
        start = ~ .x + 11,
        end = ~ .x + 21
      ),
      .names = "soc_{.fn}"
    )
  ) %>%
  # Drop missing values in birth year
  drop_na(birth_year) %>%
  # Group the data by respondent
  group_by(resp_id) %>%
  # Nest the data
  nest(soc_period = c(soc_start, soc_end)) %>%
  # Expand each respondent's socialization period
  mutate(soc_period = map(
    soc_period, ~ seq(.x$soc_start, .x$soc_end, 1)
  )) %>%
  # Unnest the data
  unnest(cols = c(soc_period)) %>%
  # Generate a vector for socialization range
  mutate(
    soc_range = str_c(
      min(soc_period),
      max(soc_period),
      sep = "-"
    )) %>%
  # Ungroup the data
  ungroup() %>%
  # Generate Socialization Age
  mutate(soc_age = soc_period - birth_year) %>%
  # Filter Socialization Age <= Current Age
  filter(soc_age <= age) %>%
  # Merge in the V-Dem Socialization Variables
  left_join(
    context_df %>%
      select(
        country:vdem_iso,
        soc_pctfemleg10 = pct_female_leg,
        soc_libdem10 = libdem,
        soc_polyarchy10 = polyarchy,
        soc_legislature10 = legislature,
        soc_election_year10 = election_year,
        soc_regime_type10 = regime_type
      ),
    by = c("country", "soc_period" = "year")
  ) 

# Recovering the ISO3 and V-Dem IDs
soc_ids_ten <- ivs_soc_df_ten %>%
  # Group the data by respondent
  group_by(resp_id) %>%
  # Collapse the IDs by respondent
  summarise(across(
    iso_code:vdem_iso,
    ~ paste0(unique(.x), collapse = ', '),
    .names = "soc_{.col}10"
  )) %>%
  # Ungroup the data
  ungroup()

ivs_soc_df_ten <- ivs_soc_df_ten %>% 
  # Group the data by identifiers
  group_by(country, year, resp_id) %>%
  # Collapse the data by respondent
  summarise(
    # Respondent's Five-Year Socialization Window
    window_10y = unique(soc_range),
    # Take the maximum value of Female HoG and Regime Type
    across(
      c(soc_legislature10:soc_election_year10, soc_regime_type10),
      ~ max(as.integer(.x), na.rm = T)
    ),
    # Take the average of continuous and count variables
    across(
      soc_pctfemleg10:soc_polyarchy10,
      ~ mean(.x, na.rm = T)
    )
  ) %>%
  # Ungroup the data
  ungroup() %>%
  # Fix any non-finite values generated by aggregation
  mutate(across(
    starts_with("soc_"),
    ~ case_when(
      is.infinite(.x) | is.nan(.x) ~ NA_real_,
      TRUE ~ .x
    )
  )) %>% 
  # Ensure no observations were duplicated
  distinct(resp_id, .keep_all = TRUE)

#------------------------------------------------------------------------------#
#---------------------------Full Socialization Data-----------------------------
#------------------------------------------------------------------------------#

ivs_soc_df_full <- soc_ids_ten %>% 
  # Merge in the socialization ids
  left_join(soc_ids_five, by = "resp_id") %>% 
  # Merge in the ten year socialization data
  left_join(ivs_soc_df_ten, by = "resp_id") %>% 
  # Merge in the five year socialization data
  left_join(ivs_soc_df_five, by = c("resp_id", "country", "year")) %>% 
  # Merge in the Contemporary V-Dem, WhoGov, and CNTS Data
  inner_join(context_df, by = c("country", "year")) %>%
  # Inner join the ivs data
  inner_join(ivs_df_sub, by = c("country", "year", "resp_id")) %>%
  # Ensure no observations were duplicated
  distinct(resp_id, .keep_all = TRUE) %>%
  # Group the data by country
  group_by(country) %>%
  # Identify the minimum year a respondent could have been socialized
  mutate(
    min_legislature_5y = min(case_when(
      soc_legislature5 == 1 ~ (birth_year + 16),
      TRUE ~ NA_real_
    ), 
    na.rm = T
    ),
    min_legislature_10y = min(case_when(
      soc_legislature10 == 1 ~ (birth_year + 11),
      TRUE ~ NA_real_
    ), 
    na.rm = T
    )
  ) %>%
  # Ungroup the data
  ungroup()

# Construct Birth Cohorts
ivs_soc_democ <- ivs_soc_df_full %>%
  # Expand each country for the period 1898 to 2007
  expand(country, birth_year = seq(1890, 2009, 1)) %>%
  # Generate five and ten year intervals for birth cohorts
  mutate(
    cohort_id_5y = rep(1:24, each = 5, length.out = n()),
    cohort_id_10y = rep(1:12, each = 10, length.out = n())
    ) %>%
  # Group the data by birth cohort
  group_by(cohort_id_5y) %>%
  # Generate a vector for cohort period
  mutate(
    cohort_5y_kk = str_c(
      min(birth_year),
      max(birth_year),
      sep = "-"
    )) %>%
  # Group the data by birth cohort
  group_by(cohort_id_10y) %>%
  # Generate a vector for cohort period
  mutate(
    cohort_10y_kk = str_c(
      min(birth_year),
      max(birth_year),
      sep = "-"
    )) %>%
  # Ungroup the data
  ungroup() %>%
  # Merge the cohort data with the main IVS file
  inner_join(ivs_soc_df_full, by = c("country", "birth_year")) %>% 
  # Arrange the data by survey and year
  .[match(ivs_soc_df_full$resp_id, .$resp_id),] %>%
  # Group the data by country
  group_by(country) %>%
  # Keep coutries that were at least electoral democracies at some point in their history
  filter(max(soc_regime_type10, na.rm = T) >= 3) %>%
  # Ungroup the data
  ungroup()

# Write the data to an RDS file
write_rds(
  x = ivs_soc_democ, 
  file = "data/IVS/IVS_Socialization_Democracies.rds",
  compress = "gz",
  compression = 9L
)
